:hammer: refactor: 구조 리팩토링
